requirements.txt
flask
google-genai
dotenv
requests